requirements.txt
javalang==0.13.0
GitPython==3.1.37
colorlog==6.7.0
SQLAlchemy==2.0.23
PyMySQL==1.1.0
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.3